fix(): Glossary filestream closing
The using before the filestream was disposing of it when going out of scope (out of the try catch). If the UploadAsync didn't access what it had to access before that the program would crash. By simply using the UploadAsync with the filename we solve this issue since the UploadAsycn will handle opening and closing the stream